+++ Vault/src/utopia/vault/nosql/factory/multi/PossiblyMultiLinkedFactory.scala
@@ -5,6 +5,7 @@ package utopia.vault.nosql.factory.multi
   * @author Mikko Hilpinen
   * @since 16.5.2020, v1.6
   */
+@deprecated("Deprecated for removal. Please inherit MultiCombiningFactory instead, or create a custom LinkedFactoryLike", "v1.18")
 trait PossiblyMultiLinkedFactory[+Parent, Child] extends MultiLinkedFactory[Parent, Child]
 {
 	override def isAlwaysLinked = false